+++ include/ORBSLAM3/KeyFrame.h
@@ -397,9 +397,9 @@ public:
 
     // objectRecognition
     void WriteToMemoryFor3DObject(
-        unsigned int &mem_pos, char *mem, const Eigen::Matrix4d &Two,
+        long long &mem_pos, char *mem, const Eigen::Matrix4d &Two,
         const bool is_superpoint);
-    unsigned int GetMemSizeFor3DObject(const bool is_superpoint);
+    long long GetMemSizeFor3DObject(const bool is_superpoint);
 
     void SetKeyPoints(std::vector<cv::KeyPoint> &keypoints);
     void SetDesps(const cv::Mat &desps);
